+++ Cims/models/CreateTechSupportItemDetails.cs
@@ -17,6 +17,7 @@ namespace Oci.CimsService.Models
 {
     /// <summary>
     /// Details about the issue that the technical support ticket relates to. Avoid entering confidential information.
+    /// For information about `TECH` support tickets, see [Creating a Technical Support Request](https://docs.cloud.oracle.com/iaas/Content/GSG/support/create-incident-technical.htm).
     /// 
     /// </summary>
     public class CreateTechSupportItemDetails : CreateItemDetails