+++ Cims/requests/ListIncidentsRequest.cs
@@ -30,7 +30,9 @@ namespace Oci.CimsService.Requests
         public string CompartmentId { get; set; }
         
         /// <value>
-        /// The Customer Support Identifier (CSI) associated with the support account.
+        /// The Customer Support Identifier (CSI) number associated with the support account.
+        /// The CSI is required for technical support tickets and optional for limits and billing tickets.
+        /// 
         /// </value>
         [Oci.Common.Http.HttpConverter(Oci.Common.Http.TargetEnum.Header, "csi")]
         public string Csi { get; set; }
@@ -75,6 +77,8 @@ namespace Oci.CimsService.Requests
         
         /// <value>
         /// User OCID for Oracle Identity Cloud Service (IDCS) users who also have a federated Oracle Cloud Infrastructure account.
+        /// User OCID is mandatory for OCI Users and optional for Multicloud users.
+        /// 
         /// </value>
         [Oci.Common.Http.HttpConverter(Oci.Common.Http.TargetEnum.Header, "ocid")]
         public string Ocid { get; set; }
@@ -111,6 +115,8 @@ namespace Oci.CimsService.Requests
         
         /// <value>
         /// The OCID of identity domain.
+        /// DomainID is mandatory if the user is part of Non Default Identity domain.
+        /// 
         /// </value>
         [Oci.Common.Http.HttpConverter(Oci.Common.Http.TargetEnum.Header, "domainid")]
         public string Domainid { get; set; }