+++ Cims/requests/UpdateIncidentRequest.cs
@@ -40,7 +40,19 @@ namespace Oci.CimsService.Requests
         public UpdateIncident UpdateIncidentDetails { get; set; }
         
         /// <value>
-        /// The Customer Support Identifier (CSI) associated with the support account.
+        /// The OCID of the tenancy.
+        /// </value>
+        /// <remarks>
+        /// Required
+        /// </remarks>
+        [Required(ErrorMessage = "CompartmentId is required.")]
+        [Oci.Common.Http.HttpConverter(Oci.Common.Http.TargetEnum.Query, "compartmentId")]
+        public string CompartmentId { get; set; }
+        
+        /// <value>
+        /// The Customer Support Identifier (CSI) number associated with the support account.
+        /// The CSI is required for technical support tickets and optional for limits and billing tickets.
+        /// 
         /// </value>
         [Oci.Common.Http.HttpConverter(Oci.Common.Http.TargetEnum.Header, "csi")]
         public string Csi { get; set; }
@@ -52,12 +64,6 @@ namespace Oci.CimsService.Requests
         public string OpcRequestId { get; set; }
         
         /// <value>
-        /// The OCID of the tenancy.
-        /// </value>
-        [Oci.Common.Http.HttpConverter(Oci.Common.Http.TargetEnum.Query, "compartmentId")]
-        public string CompartmentId { get; set; }
-        
-        /// <value>
         /// For optimistic concurrency control. In the PUT or DELETE call for a resource, set the `if-match` parameter to the value of the etag from a previous GET or POST response for that resource. The resource will be updated or deleted only if the etag you provide matches the resource's current etag value.
         /// </value>
         [Oci.Common.Http.HttpConverter(Oci.Common.Http.TargetEnum.Header, "if-match")]
@@ -65,6 +71,8 @@ namespace Oci.CimsService.Requests
         
         /// <value>
         /// User OCID for Oracle Identity Cloud Service (IDCS) users who also have a federated Oracle Cloud Infrastructure account.
+        /// User OCID is mandatory for OCI Users and optional for Multicloud users.
+        /// 
         /// </value>
         [Oci.Common.Http.HttpConverter(Oci.Common.Http.TargetEnum.Header, "ocid")]
         public string Ocid { get; set; }
@@ -95,6 +103,8 @@ namespace Oci.CimsService.Requests
         
         /// <value>
         /// The OCID of identity domain.
+        /// DomainID is mandatory if the user is part of Non Default Identity domain.
+        /// 
         /// </value>
         [Oci.Common.Http.HttpConverter(Oci.Common.Http.TargetEnum.Header, "domainid")]
         public string Domainid { get; set; }